add classchecker

+++ 275-jvm-test/src/main/java/com/njuse/jvmfinal/instructions/conversion/I2B.java
@@ -11,7 +11,7 @@ public class I2B extends NoOperandsInstruction {
      */
     @Override
     public void execute(StackFrame frame) {
-        int b=frame.getOperandStack().popInt();
+        byte b=(byte)frame.getOperandStack().popInt();
         frame.getOperandStack().pushInt(b&0xFF);
     }
 }
